pyproject.toml
[project]
name = "Contacts_API"
version = "0.1.0"
description = ""
authors = [
    {name = "Artem Bilko"}
]
readme = "README.md"
requires-python = ">=3.11.0,<4.0"
dependencies = [
    "fastapi (>=0.116.2,<0.117.0)",
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "alembic (>=1.16.5,<2.0.0)",
    "pydantic (>=2.11.9,<3.0.0)",
    "email-validator (>=2.3.0,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "bcrypt (>=4.3.0,<5.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "python-jose (>=3.5.0,<4.0.0)",
    "fastapi-mail (>=1.5.0,<2.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "slowapi (>=0.1.9,<0.2.0)",
    "cloudinary (>=1.44.1,<2.0.0)",
    "redis (>=5.0.8,<6.0.0)",
    "pytest-cov (>=7.0.0,<8.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = "-q --cov=app --cov-report=term-missing"
pythonpath = ["."]

[dependency-groups]
dev = [
    "sphinx (>=8.2.3,<9.0.0)",
    "pytest (>=8.3.3,<9.0.0)",
]
